moved struct to use chainstate::stacks::db::blocks::MessageSignatureList;
